+++ src/main/webapp/WEB-INF/views/book/book_list.jsp
