+++ src/main/webapp/WEB-INF/views/book/book_page.jsp
